requirements.txt
# 追加的方式来添加依赖，不要插入中间

aiocache==0.11.1
aiofiles==0.8.0
aioinflux==0.9.0
aioredis==2.0.0a1
aiosqlite==0.17.0
aliyun-python-sdk-core==2.13.36
aliyun-python-sdk-sts==3.1.0
asyncpg==0.25.0
cached-property==1.5.2
cachetools==4.2.2
celery==5.0.5
ciso8601==2.1.3
click==7.1.2
click-didyoumean==0.0.3
Cython==0.29.23
docopt==0.6.2
fastapi==0.65.1
fastapi-pagination==0.9.1
hestia==0.6.0
ipykernel==6.15.0
ipython==8.4.0
notebook==6.4.12
jupyterlab==3.2.3
kombu==5.0.2
kubernetes==21.7.0
logzero==1.7.0
matplotlib==3.3.4
munch==2.5.0
nbclassic==0.2.8
numpy==1.22.3
nvidia-ml-py==11.470.66
oss2==2.15.0
pandas==1.4.2
passlib[bcrypt]==1.7.4
prometheus-client==0.13.1
prometheus-fastapi-instrumentator==5.9.1
psutil==5.5.1
psycopg2-binary==2.9.2
pycurl==7.43.0.6
pytest==6.2.4
pytest-asyncio==0.15.1
pytest-forked==1.3.0
pytest-repeat==0.9.1
pytest-xdist==2.2.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.5
PyYAML==5.4.1
pyzmq==22.3.0
redis==3.5.3
rich==12.0.1
setuptools==62.6.0
sqlparams==3.0.0
sysv-ipc==1.1.0
SQLAlchemy==1.4.39
tornado==6.1
treelib==1.6.1
ujson==4.0.2
uvicorn==0.13.4
wheel==0.34.2
zope.interface==5.4.0
zope.sqlalchemy==1.4
# 20221028 追加
loguru==0.5.3
uvloop==0.16.0
schedule==1.1.0
setproctitle==1.2.2
influxdb==5.3.1
tabulate==0.8.9
wcwidth==0.2.5
swagger-ui-py==22.7.13
motor==3.0.0
pybind11==2.10.0
influxdb-client==1.33.0
# 20221209
supervisor==4.2.4
# 20230119
openapi-schema-validator==0.4.0
openpyxl==3.0.9
# 20230131
envsubst==0.1.5
pylint==2.16.2
astunparse==1.6.3
